feat(refact): Refactored Navigation code
